+++ src/core/metadata/model.ts
@@ -1,3 +1,4 @@
+import { ArrayType } from './model';
 /**
  * Copyright (c) 2017 [Arthur Xie]
  * <https://github.com/kite-js/kite>
@@ -16,6 +17,7 @@
 import { KiteError } from '../error';
 import 'reflect-metadata';
 import * as vm from 'vm';
+import { Class } from '../types/class';
 
 const MK_KITE_MODEL = 'kite:model'
 const MK_KITE_INPUTS = 'kite:inputs';
@@ -62,7 +64,7 @@ type KiteModel = {
  * `new User()` when client request comes in, without any constructor parameter, so if a Kite mode is 
  * coded like following, it'll not work as expected:
  * ```typescript
- * @Model
+ * @Model()
  * class User {
  *      @In() name: string;
  *      @In() password: string;
@@ -84,18 +86,56 @@ type KiteModel = {
  * NOTE: This expression is invalid in typescript because of type checking,
  * but, it does work in Kite at run time, no parameter is passed to constructor.
  */
-export function Model() {
-    return function (constructor: Function) {
+export function Model(globalRule?: FilterRule) {
+    return function <T extends Class>(constructor: T) {
         Reflect.defineMetadata(MK_KITE_MODEL, true, constructor);
-        createFilterFn(constructor.prototype);
+        createFilterFn(constructor.prototype, globalRule);
+
+        // return class extends constructor {
+        //     constructor(...args: any[]) {
+        //         super(args);
+        //         (this as any)[Symbol.for('isInitialized')] = true;
+        //     }
+        // };
     }
 }
 
+/**
+ * Array type definition, use only in FilterRule
+ */
+export interface ArrayType {
+    /**
+     * template of input array, etc:
+     * + Simple array: "Array<Number>", this is also the default value
+     * + Two-Dimensional array: "Array<Array<String>>"
+     * 
+     * Kite will resolve this template to filter inputs, this field is only used for
+     * describing array dimensional formats
+     */
+    template: string;
+
+    /**
+     * element type of array template, can be one of:
+     * + Number
+     * + String
+     * + Boolean
+     * + Date
+     * + Kite Model
+     * + Any other types can be set value by `new Type(input)`
+     * 
+     * if field `template` is declared clearly with the following types, this field can be omitted:
+     * + Number
+     * + String
+     * + Boolean
+     * + Date
+     */
+    elementType?: Function;
+}
 
 /**
  * Client inputs parameters filter rules
  */
-export type FilterRule = {
+export interface FilterRule {
     /**
      * type: boolean - indicates a input field is required or not, default is "false"
      * + `true` - the field is required, if it's omitted from input, Kite will throw an 
@@ -126,19 +166,35 @@ export type FilterRule = {
 
     /**
      * different behavior will be took in filter:
-     * - model property type is "String": define the minimal length of input string
+     * - model property type is "String": define the minimal value of a string, 
+     *  strings are compared based on standard lexicographical ordering, using Unicode values. example:
+     *  - 'a' - if input string < 'a' an error will be thrown
+     *  - '2017-09-01' - if input string < '2017-09-01' an error will be thrown
      * - model property type is "Number": define the minimal value of input number
      * - others: ignored
      */
-    min?: number
+    min?: number | string | Date
 
     /**
      * different behavior will be took in filter:
-     * - model property type is "String": define the maximal length of input string
+     * - model property type is "String": define the maximal value of a string
+     * strings are compared based on standard lexicographical ordering, using Unicode values. example:
+     *  - 'z' - if input string > 'z' an error will be thrown
+     *  - '2017-09-30' - if input string > '2017-09-30' an error will be thrown
      * - model property type is "Number": define the maximal value of input number
      * - others: ignored
      */
-    max?: number
+    max?: number | string | Date
+
+    /**
+     * defind the minimal length of a string, affects only on string types
+     */
+    minLen?: number
+
+    /**
+     * defind the maximal length of a string, affects only on string types
+     */
+    maxLen?: number
 
     /**
      * limit input string with special length, example:
@@ -197,16 +253,21 @@ export type FilterRule = {
     group?: string | string[];
 
     /**
-     * accept empty string, default is `false`, this field is only works when 
-     * `required` is set to `true`
+     * accept empty values, the following values are considered as empty:
+     * + `""` - an empty string
+     * + `" "` - whitespace string
+     * + `null` - null in JSON object
+     * + `[]` - an empty array
      * 
-     * `true` - allow empty string as input value
-     * `false` - disallow empty string as input value
+     * default is `false`, disallow empty values
+     * 
+     * `true` - allow empty values as input value
+     * `false` - disallow empty values as input value
      * 
      * By default, Kite filters out a string if it's empty; therefore controllers always 
      * receives nonempty strings from "String" typed parameters.
      * 
-     * In most cases, people do not want "required" fields are set by empty strings, just like 
+     * In most cases, people do not want "required" fields are empty strings, just like 
      * HTML `<input>` tag, if attribute "required" is set but nothing inputed, browser will warn you.
      * 
      * But sometimes people do, for example empty string is frequently used in database design, 
@@ -214,17 +275,36 @@ export type FilterRule = {
      * Kite accept empty strings.
      * 
      */
-    empty?: boolean;
+    allowEmpty?: boolean;
 
     /**
-     * trim string, default is 'true'.
+     * disable trim action from strings
      * 
-     * trim benginning & ending spaces from string
+     * Kite trim benginning & ending spaces from string defaultly:
      * 
-     * `true` - do trim, default value
-     * `false` - do not trim
+     * `true` - do not trim
+     * `false` - trim
      */
-    trim?: boolean;
+    noTrim?: boolean;
+
+    /**
+     * element type of property, used when source type is Array. if this filter is omitted, Kite will
+     * pass the whole input array to the controller.
+     * 
+     * example: 
+     * ```typescript
+     * class Foo {
+     *      @In({
+     *          arrayType: {
+     *              template: 'Array<Number>',
+     *              elementType: Number
+     *          }
+     *      })
+     *      arr: Array<Number>;
+     * }
+     * ```
+     */
+    arrayType?: ArrayType;
 }
 
 /**
@@ -255,9 +335,6 @@ export function In(rules: FilterRule = {}) {
     return function (target: Object, property: string) {
         let filterProperties: Map<string, FilterRule> = Reflect.getMetadata(MK_KITE_INPUTS, target) || new Map<string, FilterRule>();
 
-        // Assign default rules
-        Object.assign(rules, { trim: true });
-
         filterProperties.set(property, rules);
         // Save the rules to class metadata
         Reflect.defineMetadata(MK_KITE_INPUTS, filterProperties, target);
@@ -277,7 +354,7 @@ export function In(rules: FilterRule = {}) {
  * ```
  * @param target Target class
  */
-function createFilterFn(target: KiteModel): void {
+function createFilterFn(target: KiteModel, globalRule: FilterRule): void {
     // Keep the original _$filter function if exists
     if (target._$filter) {
         return;
@@ -298,8 +375,39 @@ function createFilterFn(target: KiteModel): void {
     let quoteRegx = /'/g;
     let groups: string[][] = [];
 
+    /**
+     * Get type name for filter function and put it to factory parameter list if type is not exist
+     * @param type 
+     */
+    function getTypeName(type: Function): String {
+        // property is other type or Kite Module, pass the original type to this filter function
+        // so it can create right objects
+        // if type is already in args, use the existing one, else create a new arg
+        let idx = args.indexOf(type);
+        let typename;
+        if (idx === -1) {
+            typename = '_' + type.name;
+            // if another type has a same name, but prototype is different, give it a new name
+            // this may happen in namespaces, two diferent namespaces both have the same object, 
+            // `type.name` only give the name
+            if (argnames.includes(typename)) {
+                typename += argnames.length;
+            }
+            argnames.push(typename);
+            args.push(type);
+        } else {
+            typename = argnames[idx];
+        }
+
+        return typename;
+    }
+
     // Loop every @In decorated properties and create filter code
     for (let [property, rule] of inputs) {
+        // if there is global rule defined, merge rules
+        if (globalRule) {
+            rule = Object.assign(globalRule, rule);
+        }
         // in case of someone named a property with sigle quotation like "it's me" cause runtime errors, following will escape it
         let name = property.replace(quoteRegx, SINGLE_QUOTE);
 
@@ -318,33 +426,6 @@ function createFilterFn(target: KiteModel): void {
             }
         }
 
-        // is required paramter
-        if (rule.required) {
-            let condition;
-            if (rule.empty) {
-                condition = `inputs['${name}'] === undefined || inputs['${name}'] === null`;
-            } else {
-                condition = `inputs['${name}'] === undefined || inputs['${name}'] === '' || inputs['${name}'] === null`;
-            }
-            // throws error is input field is not given
-            fnStack.push(`if(${condition}) {
-                throw new KiteError(1020, '${name}');
-            } else `);
-        } else {
-            fnStack.push(`if(inputs['${name}'] !== undefined)`);
-        }
-
-        // custom filters
-        if (rule.filter) {
-            // let n = $customFilters.length;
-            // $customFilters.push(rule.filter);
-            let filtername = `_f${argnames.length}`;
-            argnames.push(filtername);
-            args.push(rule.filter);
-            fnStack.push(`{ this['${name}'] = ${filtername}(inputs['${name}']);}`);
-            continue;
-        }
-
         let type = Reflect.getMetadata('design:type', target, property);
 
         // force default type to string
@@ -352,13 +433,39 @@ function createFilterFn(target: KiteModel): void {
             type = String;
         }
 
+        // field is a required parameter
+        if (rule.required) {
+            let condition = rule.allowEmpty ? `inputs['${name}'] === undefined` : `inputs['${name}'] == undefined`;
+            fnStack.push(`if(${condition}) { throw new KiteError(1020, '${name}'); } else `);
+        } else {
+            let condition = rule.allowEmpty ? `inputs['${name}'] !== undefined` : `inputs['${name}'] != undefined`;
+            fnStack.push(`if(${condition})`);
+        }
+
+        // custom filters
+        if (rule.filter) {
+            let filtername = `_f${argnames.length}`;
+            argnames.push(filtername);
+            args.push(rule.filter);
+            fnStack.push(`{ this['${name}'] = ${filtername}(inputs['${name}']);}`);
+            continue;
+        }
+
+        fnStack.push('{')
         // String check points: 
         // if defined values array, input value should in the values list
         // if defined pattern, check pattern match 
-        // if defined min, max, check for length
-        if (type.prototype === String.prototype) {   // String
-            let trim = rule.trim ? '.trim()' : '';
-            fnStack.push(`{ this['${name}'] = String(inputs['${name}'])${trim};`);
+        // if defined min, max, check for minimal & maximal values
+        // if defined minLen, maxLen, check for minimal length & maximal length
+        if (type === String) {   // String
+            let trim = rule.noTrim ? '' : '.trim()';
+            fnStack.push(`this['${name}'] = String(inputs['${name}'])${trim};`);
+
+            // if "allowEmpty" is undefined or set to false, check original input for empty string '', null 
+            if (rule.required && !rule.allowEmpty) {
+                fnStack.push(`if (this['${name}'] === '' || inputs['${name}'] === null) { throw new KiteError(1032, '${name}'); }`);
+            }
+
             // check allowed values, ignore rule.pattern, rule.min, rule.max
             if (rule.values) {
                 // let src = toSource(rule.values);
@@ -373,30 +480,36 @@ function createFilterFn(target: KiteModel): void {
                 // check for string length(limited length)
                 fnStack.push(`if(this['${name}'].length !== ${rule.len}) { throw new KiteError(1030, ['${name}',${rule.len}]); }`);
             } else {
-                // check for string length
-                if (rule.min) {
-                    fnStack.push(`if(this['${name}'].length < ${rule.min}) { throw new KiteError(1023, ['${name}',${rule.min}]); }`);
+                if (rule.minLen) {
+                    fnStack.push(`if(this['${name}'].length < ${rule.minLen}) { throw new KiteError(1023, ['${name}',${rule.minLen}]); }`);
                 }
-                if (rule.max) {
-                    fnStack.push(`if(this['${name}'].length > ${rule.max}) { throw new KiteError(1024, ['${name}',${rule.max}]); }`);
+                if (rule.maxLen) {
+                    fnStack.push(`if(this['${name}'].length > ${rule.maxLen}) { throw new KiteError(1024, ['${name}',${rule.maxLen}]); }`);
                 }
             }
 
-            fnStack.push('}');
-        } else if (type.prototype === Number.prototype) {
+            // check for string minimal & maximal value
+            if (rule.min) {
+                let quotedMin = String(rule.min).replace(quoteRegx, SINGLE_QUOTE);
+                fnStack.push(`if(this['${name}'] < '${quotedMin}') { throw new KiteError(1026, ['${name}', '${quotedMin}']); }`);
+            }
+            if (rule.max) {
+                let quotedMax = String(rule.max).replace(quoteRegx, SINGLE_QUOTE);
+                fnStack.push(`if(this['${name}'] > '${quotedMax}') { throw new KiteError(1027, ['${name}', '${quotedMax}']); }`);
+            }
+
+        } else if (type === Number) {
             // Number check points:
             // 1. parse number if input is a string
             // 2. if values is defined, input should be one of these values
             // 3. if min is defined, input should great than or equal to "min"
             // 4. if max is defined, input should less than or equal to "max"
-            fnStack.push(
-                `{
-        let num = Number(inputs['${name}']);
-        if(isNaN(num)) {
-            throw new KiteError(1025, '${name}');
-        }
-        this['${name}'] = num;
-        `);
+            fnStack.push(`let num = Number(inputs['${name}']);
+                        if(isNaN(num)) {
+                            throw new KiteError(1025, '${name}');
+                        }
+                        this['${name}'] = num;
+                        `);
             // check values
             if (rule.values) {
                 let valuesname = `_v${argnames.length}`;
@@ -411,46 +524,115 @@ function createFilterFn(target: KiteModel): void {
                     fnStack.push(`if(num > ${rule.max}) { throw new KiteError(1027, ['${name}', ${rule.max}]); } `);
                 }
             }
-            fnStack.push('}');
-        } else if (type.prototype === Boolean.prototype) {
+        } else if (type === Boolean) {
             // Boolean
             fnStack.push(`this['${name}'] = typeof inputs['${name}'] === 'string' ?
                 ['0', '', 'false'].indexOf(inputs['${name}'].toLowerCase()) === -1 :
                 Boolean(inputs['${name}']); `);
-        } else if (type.prototype === Array.prototype) {
+        } else if (type === Date) {
+            fnStack.push(`let date = new Date(inputs['${name}']);
+                if (!date.valueOf()) {
+                    throw new KiteError(1031, '${name}');
+                }`);
+
+            // validate min & max value of date
+            if (rule.min && rule.min instanceof Date) {
+                let min = rule.min.toISOString();
+                fnStack.push(`if(date < new Date('${min}')) { throw new KiteError(1026, ['${name}', '${min}']); }`);
+            }
+            if (rule.max && rule.max instanceof Date) {
+                let max = rule.max.toISOString();
+                fnStack.push(`if(date > new Date('${max}')) { throw new KiteError(1027, ['${name}', '${max}']); } `);
+            }
+
+            fnStack.push(`this['${name}'] = date;`);
+        } else if (type === Array) {
             // Array
             // [ISSUE] https://github.com/Microsoft/TypeScript/issues/7169
             // Since Typescript does not emmit array types to metadata, we don't know array element types here,
-            // so we can't parse the raw data to it's declared type
-            // TODO: add "type" to rule
-            fnStack.push(`this['${name}'] = inputs['${name}'];`);
-        } else {
-            // if type is already in args, use the existing one, else create a new arg
-            let idx = args.indexOf(type);
-            let typename;
-            if (idx === -1) {
-                typename = '_' + type.name;
-                // if another type has the same name, but type prototype is different, give a new name
-                // this may happen in namespaces, two diferent namespaces both have the same object, 
-                // `type.name` only give the name
-                if (argnames.includes(typename)) {
-                    typename += argnames.length;
-                }
-                argnames.push(typename);
-                args.push(type);
-            } else {
-                typename = argnames[idx];
+            // so we can't parse the raw data to its declared type
+
+            fnStack.push(`let input = inputs['${name}'];
+                        if (!Array.isArray(input)) {
+                            input = [input];
+                        }`);
+
+            // not allow empty array
+            if (!rule.allowEmpty) {
+                fnStack.push(`else if (!input.length) { throw new KiteError(1033, '${name}'); }`);
             }
 
-            // If it's a Kite model, create a model object and call _$filter to filter the inputs
-            if (Reflect.getMetadata(MK_KITE_MODEL, type)) {
-                fnStack.push(`{this['${name}'] = new ${typename}(); this['${name}']._$filter(inputs['${name}']);}`);
+            if (rule.arrayType) {
+                // resolve template 'Array<T>', 'Array<Array<T>>'
+                function parseArray(tpl: string) {
+                    if (tpl.startsWith('Array<') && tpl.endsWith('>')) {
+                        tpl = tpl.replace(/^Array<(.*)>$/, '$1');
+                        fnStack.push('input.map( input => ');
+                        parseArray(tpl);
+                        fnStack.push(')');
+                    } else {
+                        // if element type is not specified, determin type by template string
+                        if (!rule.arrayType.elementType) {
+                            switch (tpl.toLocaleLowerCase()) {
+                                case 'number':
+                                    rule.arrayType.elementType = Number;
+                                    break;
+                                case 'boolean':
+                                    rule.arrayType.elementType = Boolean;
+                                    break;
+                                case 'date':
+                                    rule.arrayType.elementType = Date;
+                                    break;
+                                case 'string':
+                                default:
+                                    rule.arrayType.elementType = String;
+                            }
+                        }
+
+                        let elementTypeStr;
+                        switch (rule.arrayType.elementType) {
+                            case Number:
+                                elementTypeStr = 'Number';
+                                break;
+                            case Boolean:
+                                elementTypeStr = 'Boolean';
+                                break;
+                            case String:
+                                elementTypeStr = 'String'
+                                break;
+                            case Date:
+                                elementTypeStr = 'new Date';
+                                break;
+                            default:
+                                let typename = getTypeName(rule.arrayType.elementType);
+                                if (isKiteModel(rule.arrayType.elementType)) {
+                                    elementTypeStr = `new ${typename}()._$filter`;
+                                } else {
+                                    elementTypeStr = `new ${typename}`;
+                                }
+                        }
+                        fnStack.push(`${elementTypeStr}(input)`);
+                    }
+                }
+
+                let template = rule.arrayType.template.replace(/\s/g, '');
+                fnStack.push(`this['${name}'] = `);
+                parseArray(template);
             } else {
-                // If it's not a Kite model, pass the input value to constructor and create an object
-                fnStack.push(`{ this['${name}'] = new ${typename}(typeof inputs === 'object' ? inputs['${name}'] : inputs); }`);
+                fnStack.push(`this['${name}'] = input;`);
+            }
+        } else {
+            let typename = getTypeName(type);
+            // If it is a Kite model, create a model object and call _$filter to filter the inputs
+            if (isKiteModel(type)) {
+                fnStack.push(`this['${name}'] = new ${typename}(); this['${name}']._$filter(inputs['${name}']);`);
+            } else {
+                // If it is not a Kite model, pass the input value to constructor and create an object
+                fnStack.push(`this['${name}'] = new ${typename}(typeof inputs === 'object' ? inputs['${name}'] : inputs);`);
             }
         }
 
+        fnStack.push('}');
     }
 
     // group the properties
@@ -503,7 +685,6 @@ function createFilterFn(target: KiteModel): void {
     fnStack.unshift(`(function(${argnames}) { return function(inputs) {`);   // start of function
     fnStack.push('return this;} })');  // end of return function{...}
     let fnBody = fnStack.join('\n');
-    fnStack.length = 0;
 
     let fn = vm.runInThisContext(fnBody, { filename: `__${target.constructor.name}._$filter.vm` });
     target._$filter = fn(...args);